Get rid of compiler warnings.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@14841 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/RevisionGraph/RevisionGraphDlgDraw.cpp
@@ -744,7 +744,7 @@ void CRevisionGraphWnd::DrawGraph(CDC* pDC, const CRect& rect, int nVScrollPos, 
 
     Graphics graphics (*memDC);
     graphics.SetPageUnit (UnitPixel);
-    graphics.TranslateTransform (drawRect.left, drawRect.top);
+    graphics.TranslateTransform ((REAL)drawRect.left, (REAL)drawRect.top);
 
 	if (m_fZoomFactor > 0.2f)
         DrawShadows (graphics, logRect, offset);
